tests: run test_gzip for curl_httpclient also
move simple_httpclient test_gzip to the shared httpclient tests,
to test the decompress_response option for curl_httpclient as well
